requirements.txt

pandas>=2.0.0
numpy>=1.24.0
jiwer>=3.0.0
soundfile>=0.12.1
librosa>=0.10.1
openai-whisper>=20240930
faster-whisper>=1.0.0
vosk>=0.3.45
transformers>=4.44.0
torch>=2.3.1
accelerate>=0.33.0
huggingface_hub>=0.23.0
requests>=2.31.0
sounddevice
kenlm
pyctcdecode
